projeto_final/requirements.txt
autopep8==1.6.0
numpy==1.22.1
pycodestyle==2.8.0
PyOpenGL==3.1.5
toml==0.10.2
